wip: validation hook
