+++ src/components/exchanges/universal/clipboardViewer/ClipboardViewer.tsx
@@ -1,8 +1,8 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React, { useState, useEffect } from 'react';
 import { ClipboardIcon, ClipboardText, Container, ContentContainer, TextContainer, Title } from './styles';
 import { useTranslation } from 'react-i18next';
 import { readText } from '@tauri-apps/plugin-clipboard-manager';
-import { invoke } from '@tauri-apps/api/core';
+import { useValidate } from '@app/hooks/wallet/useValidate.ts';
 
 export interface ClipboardViewerProps {
     handlePaste: (text: string) => void;
@@ -12,14 +12,9 @@ export const ClipboardViewer: React.FC<ClipboardViewerProps> = ({ handlePaste })
     const [clipboardText, setClipboardText] = useState<string | undefined>();
     const { t } = useTranslation('exchange');
 
-    const validateAddress = useCallback(async (value: string) => {
-        try {
-            await invoke('verify_address_for_send', { address: value });
-            return true;
-        } catch (_) {
-            return false;
-        }
-    }, []);
+    const { validateAddress, validationError } = useValidate();
+
+    console.debug(`validationError= `, validationError);
 
     useEffect(() => {
         const readClipboard = async () => {
